refactor(simulation): split core/index.ts and extract world.ts helpers
core/index.ts (1664 lines) → 5 files:
- virtual-time.ts: VirtualTime + timer scheduling
- memory-storage.ts: InMemoryStorage + filesystem helpers
- runtime-doubles.ts: InMemoryPaths, SequentialIds, SealedEnv
- mock-process.ts: MockProcessSpawner, MockChildProcess, MockDurableTransport
- index.ts: SimulationRuntime composition root + re-exports

world.ts (929 lines) → 4 files:
- no-real-io.ts: NoRealIo monitor (global intercept lifecycle)
- scenario-http.ts: ScenarioHttpRequest/Response doubles
- normalize.ts: WorldConfig normalization functions
- world.ts: SimulationWorld class only (~500 lines)

Pure file reorganization — no behavioral changes.
